requirements.txt
aiogram>=3.0.0b3
aioredis~=2.0
environs~=9.0
psycopg2~=2.9.3
requests
numpy
matmatplotlib
